:recycle: refactor: renew 이벤트 명 변경
