+++ src/main/java/com/mulmeong/batchserver/comment/application/FeedCommentService.java
@@ -1,11 +1,11 @@
 package com.mulmeong.batchserver.comment.application;
 
-import com.mulmeong.event.utility.consume.DislikesCreateEvent;
-import com.mulmeong.event.utility.consume.LikesCreateEvent;
+import com.mulmeong.event.utility.consume.DislikeRenewCreateEvent;
+import com.mulmeong.event.utility.consume.LikeRenewCreateEvent;
 
 public interface FeedCommentService {
 
-    void likeCountRenew(LikesCreateEvent message);
+    void likeCountRenew(LikeRenewCreateEvent message);
 
-    void dislikeCountRenew(DislikesCreateEvent message);
+    void dislikeCountRenew(DislikeRenewCreateEvent message);
 }
