+++ src/main/java/com/mulmeong/batchserver/comment/application/FeedRecommentServiceImpl.java
@@ -4,8 +4,8 @@ import com.mulmeong.batchserver.comment.domain.document.FeedRecomment;
 import com.mulmeong.batchserver.comment.infrastructure.repository.FeedRecommentReadRepository;
 import com.mulmeong.batchserver.utility.infrastructure.repository.DislikesRepository;
 import com.mulmeong.batchserver.utility.infrastructure.repository.LikesRepository;
-import com.mulmeong.event.utility.consume.DislikesCreateEvent;
-import com.mulmeong.event.utility.consume.LikesCreateEvent;
+import com.mulmeong.event.utility.consume.DislikeRenewCreateEvent;
+import com.mulmeong.event.utility.consume.LikeRenewCreateEvent;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
@@ -20,7 +20,7 @@ public class FeedRecommentServiceImpl implements FeedRecommentService {
     private final DislikesRepository dislikesRepository;
 
     @Override
-    public void likeCountRenew(LikesCreateEvent message) {
+    public void likeCountRenew(LikeRenewCreateEvent message) {
 
         FeedRecomment feedRecommentReadUpdate = feedRecommentReadRepository
                 .findByRecommentUuid(message.getKindUuid()).orElseThrow();
@@ -31,7 +31,7 @@ public class FeedRecommentServiceImpl implements FeedRecommentService {
     }
 
     @Override
-    public void dislikeCountRenew(DislikesCreateEvent message) {
+    public void dislikeCountRenew(DislikeRenewCreateEvent message) {
 
         FeedRecomment feedRecommentReadUpdate = feedRecommentReadRepository
                 .findByRecommentUuid(message.getKindUuid()).orElseThrow();
