+++ src/main/java/com/mulmeong/batchserver/comment/application/ShortsCommentService.java
@@ -1,12 +1,12 @@
 package com.mulmeong.batchserver.comment.application;
 
-import com.mulmeong.event.utility.consume.DislikesCreateEvent;
-import com.mulmeong.event.utility.consume.LikesCreateEvent;
+import com.mulmeong.event.utility.consume.DislikeRenewCreateEvent;
+import com.mulmeong.event.utility.consume.LikeRenewCreateEvent;
 
 public interface ShortsCommentService {
 
-    void likeCountRenew(LikesCreateEvent message);
+    void likeCountRenew(LikeRenewCreateEvent message);
 
-    void dislikeCountRenew(DislikesCreateEvent message);
+    void dislikeCountRenew(DislikeRenewCreateEvent message);
 
 }
