+++ src/main/java/com/mulmeong/batchserver/comment/config/ShortsCommentBatchConfig.java
@@ -31,7 +31,7 @@ import org.springframework.transaction.PlatformTransactionManager;
 
 @Configuration
 @EnableScheduling
-@EnableBatchProcessing(dataSourceRef = "contestDataSource", transactionManagerRef = "contestTransactionManager")
+@EnableBatchProcessing
 @RequiredArgsConstructor
 @Slf4j
 public class ShortsCommentBatchConfig {
