+++ src/main/java/com/mulmeong/batchserver/contest/config/ContestBatchConfig.java
@@ -31,13 +31,12 @@ import java.util.stream.Collectors;
 
 @Configuration
 @EnableScheduling
-@EnableBatchProcessing(dataSourceRef = "contestDataSource", transactionManagerRef = "contestTransactionManager")
+@EnableBatchProcessing
 @RequiredArgsConstructor
 @Slf4j
 public class ContestBatchConfig {
     private final ContestRepository contestRepository;
     private final RedisTemplate<String, String> redisTemplate;
-    @Qualifier("contestTransactionManager")
     private final PlatformTransactionManager transactionManager;
     private final ContestKafkaPublisher contestKafkaPublisher;
     private static final String VOTE_COUNT_KEY = "contest:%d:post:votes";
