+++ src/main/java/com/mulmeong/batchserver/shorts/application/ShortsService.java
@@ -1,10 +1,10 @@
 package com.mulmeong.batchserver.shorts.application;
 
-import com.mulmeong.event.utility.consume.DislikesCreateEvent;
-import com.mulmeong.event.utility.consume.LikesCreateEvent;
+import com.mulmeong.event.utility.consume.DislikeRenewCreateEvent;
+import com.mulmeong.event.utility.consume.LikeRenewCreateEvent;
 
 public interface ShortsService {
-    void likeCountRenew(LikesCreateEvent message);
+    void likeCountRenew(LikeRenewCreateEvent message);
 
-    void dislikeCountRenew(DislikesCreateEvent message);
+    void dislikeCountRenew(DislikeRenewCreateEvent message);
 }
