+++ src/main/java/com/mulmeong/batchserver/shorts/application/ShortsServiceImpl.java
@@ -4,8 +4,8 @@ import com.mulmeong.batchserver.shorts.domain.document.ShortsRead;
 import com.mulmeong.batchserver.shorts.infrastructure.repository.ShortsReadRepository;
 import com.mulmeong.batchserver.utility.infrastructure.repository.DislikesRepository;
 import com.mulmeong.batchserver.utility.infrastructure.repository.LikesRepository;
-import com.mulmeong.event.utility.consume.DislikesCreateEvent;
-import com.mulmeong.event.utility.consume.LikesCreateEvent;
+import com.mulmeong.event.utility.consume.DislikeRenewCreateEvent;
+import com.mulmeong.event.utility.consume.LikeRenewCreateEvent;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
@@ -20,7 +20,7 @@ public class ShortsServiceImpl implements ShortsService {
     private final DislikesRepository dislikesRepository;
 
     @Override
-    public void likeCountRenew(LikesCreateEvent message) {
+    public void likeCountRenew(LikeRenewCreateEvent message) {
 
         ShortsRead shortsReadUpdate = shortsReadRepository.findByShortsUuid(message.getKindUuid()).orElseThrow();
         Long count = likesRepository.countByKindAndKindUuidAndStatus(message.getKind(), message.getKindUuid(), true);
@@ -30,7 +30,7 @@ public class ShortsServiceImpl implements ShortsService {
     }
 
     @Override
-    public void dislikeCountRenew(DislikesCreateEvent message) {
+    public void dislikeCountRenew(DislikeRenewCreateEvent message) {
         ShortsRead shortsReadUpdate = shortsReadRepository.findByShortsUuid(message.getKindUuid()).orElseThrow();
         Long count = dislikesRepository.countByKindAndKindUuidAndStatus(message.getKind(), message.getKindUuid(), true);
         log.info("count: {}", count);
